+++ src/controllers/wallet.controller.ts
@@ -1,7 +1,10 @@
 import httpStatusCodes from 'http-status-codes';
 import logger from '../config/logger';
-import { RESPONSE_ERROR, WALLET_ERROR } from '../constants/errors';
-import { WITHDRAWAL_RESPONSE } from '../constants/successMessages';
+import { ERRORS, RESPONSE_ERROR, WALLET_ERROR } from '../constants/errors';
+import {
+  REFUND_RESPONSE,
+  WITHDRAWAL_RESPONSE,
+} from '../constants/successMessages';
 import Utility from '../constants/utility';
 import WalletService from '../services/wallet.service';
 import apiResponse from '../utilities/apiResponse';
@@ -17,7 +20,7 @@ export class WalletController {
         httpStatusCodes.OK
       );
     } catch (e) {
-      logger.error('[walletController.viewBillingsByFilter]:' + e.message);
+      logger.error('[WalletController.viewBillingsByFilter]:' + e.message);
       return apiResponse.error(res, httpStatusCodes.BAD_REQUEST, {
         message: e.message,
       });
@@ -37,7 +40,7 @@ export class WalletController {
         httpStatusCodes.OK
       );
     } catch (e) {
-      logger.error('[walletController.viewListOfWallets]:' + e.message);
+      logger.error('[WalletController.viewListOfWallets]:' + e.message);
       return apiResponse.error(res, httpStatusCodes.INTERNAL_SERVER_ERROR, {
         message: RESPONSE_ERROR.RES_ERROR,
       });
@@ -60,11 +63,59 @@ export class WalletController {
         httpStatusCodes.OK
       );
     } catch (e) {
-      logger.error('[walletController.viewWallet]:' + e.message);
+      logger.error('[WalletController.viewWallet]:' + e.message);
       return Utility.apiErrorResponse(res, e, [WALLET_ERROR.UNAUTH_WALLET]);
     }
   }
 
+  // ============================== Refund ==============================
+  public static async cancelRefundRequest(req, res) {
+    try {
+      const { refundRequestId } = req.params;
+      const { user } = req;
+      await WalletService.cancelRefundRequest(refundRequestId, user.accountId);
+      return apiResponse.result(
+        res,
+        { message: REFUND_RESPONSE.REQUEST_CANCEL },
+        httpStatusCodes.OK
+      );
+    } catch (e) {
+      logger.error('[WalletController.requestRefund]:' + e.message);
+      return Utility.apiErrorResponse(res, e, [
+        ERRORS.STUDENT_DOES_NOT_EXIST,
+        WALLET_ERROR.REFUNDED,
+        WALLET_ERROR.EXISTING_REFUND,
+        WALLET_ERROR.REFUND_PERIOD_OVER,
+      ]);
+    }
+  }
+
+  public static async requestRefund(req, res) {
+    try {
+      const { contractId, contractType } = req.query;
+      const { user } = req;
+      const refundRequest = await WalletService.requestRefund(
+        contractId,
+        contractType,
+        user.accountId
+      );
+      return apiResponse.result(
+        res,
+        { message: REFUND_RESPONSE.REQUEST_CREATE, refundRequest },
+        httpStatusCodes.OK
+      );
+    } catch (e) {
+      logger.error('[WalletController.requestRefund]:' + e.message);
+      return Utility.apiErrorResponse(res, e, [
+        ERRORS.STUDENT_DOES_NOT_EXIST,
+        WALLET_ERROR.REFUNDED,
+        WALLET_ERROR.EXISTING_REFUND,
+        WALLET_ERROR.REFUND_PERIOD_OVER,
+      ]);
+    }
+  }
+
+  // ============================== Withdrawal ==============================
   public static async withdrawBalance(req, res) {
     try {
       const { user } = req;
@@ -79,7 +130,7 @@ export class WalletController {
         httpStatusCodes.OK
       );
     } catch (e) {
-      logger.error('[walletController.withdrawBalance]:' + e.message);
+      logger.error('[WalletController.withdrawBalance]:' + e.message);
       return Utility.apiErrorResponse(res, e, [
         WALLET_ERROR.UNAUTH_WALLET,
         WALLET_ERROR.NO_MONEY,
@@ -97,7 +148,7 @@ export class WalletController {
         httpStatusCodes.OK
       );
     } catch (e) {
-      logger.error('[walletController.manualChronjob]:' + e.message);
+      logger.error('[WalletController.manualChronjob]:' + e.message);
       return apiResponse.error(res, httpStatusCodes.INTERNAL_SERVER_ERROR, {
         message: e.message,
       });
