+++ src/routes/paypal.route.ts
@@ -5,7 +5,6 @@ import { PaypalController } from '../controllers/paypal.controller';
 import { requireStudent } from '../middlewares/authenticationMiddleware';
 import cart from './schema/cart.schema';
 import paypal from './schema/paypal.schema';
-import wallet from './schema/wallet.schema';
 
 const router = express.Router();
 
@@ -40,21 +39,5 @@ router.get(
   requireStudent,
   schemaValidator.params(paypal.paymentIdP),
   Utility.asyncHandler(PaypalController.viewOrder)
-);
-
-router.post(
-  '/refund',
-  passport.authenticate('isAuthenticated', { session: false }),
-  requireStudent,
-  schemaValidator.query(wallet.refundRequestQ),
-  Utility.asyncHandler(PaypalController.requestRefund)
-);
-
-router.delete(
-  '/refund/:refundRequestId',
-  passport.authenticate('isAuthenticated', { session: false }),
-  requireStudent,
-  schemaValidator.params(wallet.refundRequestIdP),
-  Utility.asyncHandler(PaypalController.cancelRefundRequest)
 );
 export default router;
