+++ src/routes/wallet.route.ts
@@ -5,6 +5,7 @@ import { WalletController } from '../controllers/wallet.controller';
 import {
   requireFinanceIfAdmin,
   requireSensei,
+  requireStudent,
 } from '../middlewares/authenticationMiddleware';
 import wallet from './schema/wallet.schema';
 
@@ -37,6 +38,22 @@ router.get(
   requireFinanceIfAdmin,
   schemaValidator.query(wallet.billingFilterQ),
   Utility.asyncHandler(WalletController.viewBillingsByFilter)
+);
+
+router.post(
+  '/refund',
+  passport.authenticate('isAuthenticated', { session: false }),
+  requireStudent,
+  schemaValidator.query(wallet.refundRequestQ),
+  Utility.asyncHandler(WalletController.requestRefund)
+);
+
+router.delete(
+  '/refund/:refundRequestId',
+  passport.authenticate('isAuthenticated', { session: false }),
+  requireStudent,
+  schemaValidator.params(wallet.refundRequestIdP),
+  Utility.asyncHandler(WalletController.cancelRefundRequest)
 );
 
 // Interim trigger to update billings to confirmed if date has passed
